compiling again

+++ lib/src/net/download_manager/download_manager.dart
@@ -8,11 +8,13 @@
 import "dart:async";
 import "dart:io";
 
+import "package:cached_network_image/cached_network_image.dart";
 import "package:dio/dio.dart";
 import "package:flutter/material.dart";
 import "package:flutter_animate/flutter_animate.dart";
 import "package:gallery/src/db/services/services.dart";
 import "package:gallery/src/db/tags/post_tags.dart";
+import "package:gallery/src/interfaces/cell/cell.dart";
 import "package:gallery/src/interfaces/logging/logging.dart";
 import "package:gallery/src/plugs/download_movers.dart";
 import "package:gallery/src/plugs/notifications.dart";
@@ -23,32 +25,80 @@ part "download_entry.dart";
 part "download_status.dart";
 part "download_handle.dart";
 
-class DownloadManager with _StatisticsTimer {
+mixin _SourceImpl on SourceStorage<DownloadHandle>
+    implements ResourceSource<DownloadHandle> {
+  @override
+  SourceStorage<DownloadHandle> get backingStorage => this;
+
+  @override
+  Future<int> clearRefresh() => Future.value(count);
+
+  @override
+  void destroy() {}
+
+  @override
+  DownloadHandle? forIdx(int idx) => get(idx);
+
+  @override
+  DownloadHandle forIdxUnsafe(int idx) => this[idx];
+
+  @override
+  Future<int> next() => Future.value(count);
+}
+
+class DownloadManager extends SourceStorage<DownloadHandle>
+    with _StatisticsTimer, _SourceImpl {
   DownloadManager(this.moverPlug, this._db);
+
+  factory DownloadManager.of(BuildContext context) =>
+      DatabaseConnectionNotifier.downloadManagerOf(context);
 
   int _notificationId = 0;
 
   int _inWork = 0;
   final _client = Dio();
-  final StreamController<void> _events = StreamController();
+  final StreamController<int> _events = StreamController();
 
   final DownloadFileService _db;
 
   final NotificationPlug notificationPlug = chooseNotificationPlug();
   final DownloadMoverPlug moverPlug;
 
-  final Map<String, _DownloadEntry> _aliveEntries = {};
+  final Map<String, int> _entriesMap = {};
+  final List<_DownloadEntry> _aliveEntries = [];
 
   static const _log = LogTarget.downloader;
   static const int maximum = 6;
 
-  StreamSubscription<void> watch(void Function(void) f) =>
-      _events.stream.listen(f);
+  // ValueKey<int> get widgetKey => ValueKey(_notificationId);
+  // Iterable<DownloadHandle> get handles => _aliveEntries.values;
 
-  ValueKey<int> get widgetKey => ValueKey(_notificationId);
-  Iterable<DownloadHandle> get handles => _aliveEntries.values;
+  @override
+  int get count => _aliveEntries.length;
 
-  factory DownloadManager.of(BuildContext context) {}
+  @override
+  DownloadHandle? get(int idx) =>
+      idx >= _aliveEntries.length ? null : _aliveEntries[0];
+
+  @override
+  Iterator<DownloadHandle> get iterator => _aliveEntries.iterator;
+
+  @override
+  DownloadHandle operator [](int index) => _aliveEntries[0];
+
+  @override
+  void operator []=(int index, DownloadHandle value) {}
+
+  @override
+  void add(DownloadHandle e, [bool silent = false]) =>
+      throw "should not be called";
+
+  @override
+  void addAll(List<DownloadHandle> l, [bool silent = false]) =>
+      throw "should not be called";
+
+  @override
+  void removeAll(List<int> idx) => throw "should not be called";
 
   // String downloadDescription(DownloadFile f) {
   //   if (_hasCancelKey(f.url)) {
@@ -141,15 +191,18 @@ class DownloadManager with _StatisticsTimer {
     _events.close();
   }
 
-  void removeAll() {
+  @override
+  void clear() {
     _aliveEntries
-      ..values.map((e) {
+      ..map((e) {
         e.token.cancel();
         e.watcher?.close();
       })
       ..clear();
+    _entriesMap.clear();
 
     _db.clear();
+    _events.add(count);
   }
 
   void remove(List<DownloadHandle> l) {
@@ -161,14 +214,16 @@ class DownloadManager with _StatisticsTimer {
 
     for (final e in l) {
       e.cancel();
-      final r = _aliveEntries.remove(e.key);
-      if (r != null) {
+      final id = _entriesMap.remove(e.key);
+      if (id != null) {
+        final r = _aliveEntries.removeAt(id);
         removed.add(r.key);
         r.watcher?.close();
       }
     }
 
     _db.deleteAll(removed);
+    _events.add(count);
   }
 
   void addLocalTags(
@@ -177,14 +232,14 @@ class DownloadManager with _StatisticsTimer {
     PostTags postTags,
   ) {}
 
-  void addAll(Iterable<DownloadEntry> downloads, SettingsData settings) {
+  void putAll(Iterable<DownloadEntry> downloads, SettingsData settings) {
     if (settings.path.isEmpty) {
       return;
     }
 
     final toDownload = downloads
         .skipWhile(
-          (element) => _aliveEntries.containsKey(element.url),
+          (element) => _entriesMap.containsKey(element.url),
         )
         .toList();
     if (downloads.isEmpty) {
@@ -197,10 +252,12 @@ class DownloadManager with _StatisticsTimer {
       if (_inWork > maximum) {
         final hold = e._copyStatus(DownloadStatus.onHold);
 
-        _aliveEntries[e.url] = _DownloadEntry(data: hold, token: CancelToken());
+        _aliveEntries.add(_DownloadEntry(data: hold, token: CancelToken()));
+        _entriesMap[e.url] = _aliveEntries.length - 1;
         toSaveDb.add(hold._toDb());
       } else {
-        _aliveEntries[e.url] = _DownloadEntry(data: e, token: CancelToken());
+        _aliveEntries.add(_DownloadEntry(data: e, token: CancelToken()));
+        _entriesMap[e.url] = _aliveEntries.length - 1;
         toSaveDb.add(e._toDb());
 
         _inWork += 1;
@@ -210,6 +267,7 @@ class DownloadManager with _StatisticsTimer {
     }
 
     _db.saveAll(toSaveDb);
+    // _events.add(null);
   }
 
   void _tryAddNew() {
@@ -222,16 +280,19 @@ class DownloadManager with _StatisticsTimer {
       return;
     }
 
-    final entries = _aliveEntries.values
+    final entries = _aliveEntries.indexed
         .where(
-          (element) => element.data.status == DownloadStatus.onHold,
+          (element) => element.$2.data.status == DownloadStatus.onHold,
         )
         .take(newAdd)
         .map(
-          (e) => _DownloadEntry(
-            data: e.data._copyStatus(DownloadStatus.inProgress),
-            token: CancelToken(),
-            watcher: e.watcher,
+          (e) => (
+            e.$1,
+            _DownloadEntry(
+              data: e.$2.data._copyStatus(DownloadStatus.inProgress),
+              token: CancelToken(),
+              watcher: e.$2.watcher,
+            )
           ),
         )
         .toList();
@@ -242,18 +303,14 @@ class DownloadManager with _StatisticsTimer {
 
     _inWork += entries.length;
 
-    _aliveEntries.addEntries(
-      entries.map(
-        (e) => MapEntry(
-          e.key,
-          e,
-        ),
-      ),
-    );
-    _db.saveAll(entries.map((e) => e.data._toDb()).toList());
+    for (final (id, e) in entries) {
+      _aliveEntries[id] = e;
+    }
+
+    _db.saveAll(entries.map((e) => e.$2.data._toDb()).toList());
 
     for (final e in entries) {
-      _download(e.key);
+      _download(e.$2.key);
     }
   }
 
@@ -263,7 +320,7 @@ class DownloadManager with _StatisticsTimer {
     required int count,
     required int total,
   }) {
-    if (count == total || !_aliveEntries.containsKey(url)) {
+    if (count == total || !_entriesMap.containsKey(url)) {
       notif.done();
       return;
     }
@@ -273,30 +330,33 @@ class DownloadManager with _StatisticsTimer {
   }
 
   void _remove(String url) {
-    _aliveEntries.remove(url)?.watcher?.close();
-    _db.deleteAll([url]);
+    final id = _entriesMap.remove(url);
+    if (id != null) {
+      _aliveEntries.removeAt(id).watcher?.close();
+      _db.deleteAll([url]);
+    }
   }
 
   void _complete(String url) {
     _remove(url);
-    _events.add(null);
+    _events.add(count);
   }
 
   void _failed(_DownloadEntry entry) {
     final newEntry = entry.data._copyStatus(DownloadStatus.failed);
 
-    _aliveEntries[entry.key] = _DownloadEntry(
+    _aliveEntries[_entriesMap[entry.key]!] = _DownloadEntry(
       data: newEntry,
       token: CancelToken(),
       watcher: entry.watcher,
     );
     _db.saveAll([newEntry._toDb()]);
 
-    _events.add(null);
+    _events.add(count);
   }
 
   Future<void> _download(String url) async {
-    final entry = _aliveEntries[url]!;
+    final entry = _aliveEntries[_entriesMap[url]!];
 
     try {
       final dir = await _tryCreateInternalDownloadDir(entry);
@@ -386,6 +446,10 @@ class DownloadManager with _StatisticsTimer {
       ),
     );
   }
+
+  @override
+  StreamSubscription<int> watch(void Function(int) f) =>
+      _events.stream.listen(f);
 
   // Future<void> _removeTempContentsDownloads() async {
   //   try {
